requirements.txt
wheel
beautifulsoup4==4.9.3
requests==2.25.1
bs4
Pillow
PyMuPDF
